+++ tests/common.py
@@ -1,14 +1,25 @@
 import asyncio
 import functools
-import pathlib
 import logging
+import pathlib
 import shutil
 import socket
+import ssl
 
 import nose
+import trustme
 
 import aioftp
 
+
+ca = trustme.CA()
+server_cert = ca.issue_server_cert("127.0.0.1", "::1")
+
+ssl_server = ssl.create_default_context(ssl.Purpose.CLIENT_AUTH)
+server_cert.configure_cert(ssl_server)
+
+ssl_client = ssl.create_default_context(ssl.Purpose.SERVER_AUTH)
+ca.configure_trust(ssl_client)
 
 PORT = 8888
 
@@ -23,7 +34,7 @@ def aioftp_setup(*, server_args=([], {}), client_args=([], {})):
             s_args, s_kwargs = server_args
             c_args, c_kwargs = client_args
 
-            def run_in_loop(s_args, s_kwargs, c_args, c_kwargs):
+            def run_in_loop(s_args, s_kwargs, c_args, c_kwargs, s_ssl=None, c_ssl=None):
                 logging.basicConfig(
                     level=logging.INFO,
                     format="%(asctime)s [%(name)s] %(message)s",
@@ -31,8 +42,8 @@ def aioftp_setup(*, server_args=([], {}), client_args=([], {})):
                 )
                 loop = asyncio.new_event_loop()
                 asyncio.set_event_loop(None)
-                server = aioftp.Server(*s_args, loop=loop, **s_kwargs)
-                client = aioftp.Client(*c_args, loop=loop, **c_kwargs)
+                server = aioftp.Server(*s_args, loop=loop, ssl=s_ssl, **s_kwargs)
+                client = aioftp.Client(*c_args, loop=loop, ssl=c_ssl, **c_kwargs)
                 try:
                     loop.run_until_complete(f(loop, client, server))
                 finally:
@@ -46,8 +57,10 @@ def aioftp_setup(*, server_args=([], {}), client_args=([], {})):
                 for factory in (aioftp.PathIO, aioftp.AsyncPathIO):
                     s_kwargs["path_io_factory"] = factory
                     run_in_loop(s_args, s_kwargs, c_args, c_kwargs)
+                    run_in_loop(s_args, s_kwargs, c_args, c_kwargs, ssl_server, ssl_client)
             else:
                 run_in_loop(s_args, s_kwargs, c_args, c_kwargs)
+                run_in_loop(s_args, s_kwargs, c_args, c_kwargs, ssl_server, ssl_client)
 
         return wrapper